fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,28 +7,27 @@
 int main(void)
 {
 	f(50);
-
 	return (0);
 }
 
 void f(int n)
 {
-	long n1 = 0;
+	unsigned long long n1 = 0;
 
-	long n2 = 1;
+	unsigned long long n2 = 1;
 
 	int i;
 
-	printf("%ld, ", n1);
+	printf("%llu, ", n1);
 
 	for (i = 0; i < n - 1; i++)
 	{
 
 		if (i == n - 2)
-			printf("%ld", n2);
+			printf("%llu", n2);
 		else
-			printf("%ld, ", n2);
-		long n3 = n1 + n2;
+			printf("%llu, ", n2);
+		unsigned long long n3 = n1 + n2;
 
 		n1 = n2;
 		n2 = n3;
